requirements.txt
numpy
pandas
torch
pillow
opencv-python
matplotlib
pytesseract
sympy
scipy
